// backend/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Side = {
  ASK: 'ASK',
  BID: 'BID'
} as const

export type Side = (typeof Side)[keyof typeof Side]


export const Type = {
  LIMIT: 'LIMIT',
  MARKET: 'MARKET'
} as const

export type Type = (typeof Type)[keyof typeof Type]


export const Status = {
  FILLED: 'FILLED',
  PARTIAL: 'PARTIAL',
  EMPTY: 'EMPTY'
} as const

export type Status = (typeof Status)[keyof typeof Status]